Added air drag for falling and jet-pack usage. Also moved spawn position back to player.js and changed the location by importing it into the environment.js
